Various portability fixes for Slinky (#477)
Various fixes to improve portability to targets/compilers that we hadn't
tested on (much):
- index_t needed some love to avoid multiple issues:
- on 32-bit systems we really want it to be a 32-bit type for
efficiency; there were various templated functions that assumed it was
always an int64_t that needed to use an `index_t` (or `T`) as the
temporary instead
- defining `index_t` as `ptrdiff_t` or `intptr_t` seems like the right
thing but introduces pain on some compilers (eg Emscripten, AppleClang)
where the result is a `long` type that is distinct from both `int32_t`
and `int64_t`, making it painful to get the overloaded ctors for `expr`
correct without an ugly mass of ifdefs. Instead, I chose to use
`std::conditional` to make it explicitly `int32_t` or `int64_t` to dodge
that particular bullet.
- expr() needs a size_t ctor; note that size_t can be a bespoke type
that doesn't match uint32_t or uint64_t
- Visual C++ fixes (alloca -> __builtin_alloca; implicit bool -> expr
conversion doesn't match gcc/clang in some cases, requiring explicit
conversion)

+++ runtime/expr.h
@@ -194,7 +194,7 @@ public:
   // Make a new constant expression.
   expr(std::int64_t x);
   expr(std::int32_t x) : expr(static_cast<std::int64_t>(x)) {}
-  expr(std::uint64_t x) : expr(static_cast<std::int64_t>(x)) {}
+  expr(std::size_t x) : expr(static_cast<std::int64_t>(x)) {}
   expr(var sym);
 
   // Make an `expr` referencing an existing node.
